出库单创建反馈信息

+++ warehouse-operation-api/src/main/java/com/baozun/scm/primservice/whinterface/model/outbound/WmsOutBoundInvoiceLineConfirm.java
@@ -30,11 +30,11 @@ public class WmsOutBoundInvoiceLineConfirm implements Serializable {
     /** 行号 */
     private String linenum;
     /** 数量 */
-    private Long qty;
+    private Double qty;
     /** 单价 */
-    private Long unitPrice;
+    private Double unitPrice;
     /** 总金额 */
-    private Long amt;
+    private Double amt;
     /** 类别 */
     private String item;
 
@@ -47,27 +47,27 @@ public class WmsOutBoundInvoiceLineConfirm implements Serializable {
         this.linenum = linenum;
     }
 
-    public Long getQty() {
+    public Double getQty() {
         return qty;
     }
 
-    public void setQty(Long qty) {
+    public void setQty(Double qty) {
         this.qty = qty;
     }
 
-    public Long getUnitPrice() {
+    public Double getUnitPrice() {
         return unitPrice;
     }
 
-    public void setUnitPrice(Long unitPrice) {
+    public void setUnitPrice(Double unitPrice) {
         this.unitPrice = unitPrice;
     }
 
-    public Long getAmt() {
+    public Double getAmt() {
         return amt;
     }
 
-    public void setAmt(Long amt) {
+    public void setAmt(Double amt) {
         this.amt = amt;
     }
